refactor(frontend): remove fetch.utils
