+++ frontend/app/.server/auth/raoidc.service.ts
@@ -5,8 +5,8 @@ import { subtle } from 'node:crypto';
 import type { ServerConfig } from '~/.server/configs';
 import { TYPES } from '~/.server/constants';
 import type { LogFactory, Logger } from '~/.server/factories';
+import type { FetchFn, FetchService } from '~/.server/http';
 import { generateCryptoKey, generateJwkId } from '~/.server/utils/crypto.utils';
-import { FetchFn, getFetchFn } from '~/.server/utils/fetch.utils';
 import type { ClientMetadata, IdToken, UserinfoToken } from '~/.server/utils/raoidc.utils';
 import { fetchAccessToken, fetchServerMetadata, fetchUserInfo, generateAuthorizationRequest, generateCodeChallenge, generateRandomState } from '~/.server/utils/raoidc.utils';
 import { expandTemplate } from '~/utils/string-utils';
@@ -103,9 +103,10 @@ export class DefaultRaoidcService implements RaoidcService {
   constructor(
     @inject(TYPES.factories.LogFactory) logFactory: LogFactory,
     @inject(TYPES.configs.ServerConfig) private readonly serverConfig: Pick<ServerConfig, 'AUTH_RAOIDC_BASE_URL' | 'AUTH_RAOIDC_CLIENT_ID' | 'AUTH_RAOIDC_METADATA_CACHE_TTL_SECONDS' | 'AUTH_JWT_PRIVATE_KEY' | 'AUTH_LOGOUT_REDIRECT_URL' | 'HTTP_PROXY_URL'>,
+    @inject(TYPES.http.FetchService) fetchService: FetchService,
   ) {
     this.log = logFactory.createLogger('DefaultRaoidcService');
-    this.fetchFn = getFetchFn(serverConfig.HTTP_PROXY_URL);
+    this.fetchFn = fetchService.getFetchFn({ proxyUrl: serverConfig.HTTP_PROXY_URL });
 
     // Configure caching for raoidc operations
     this.fetchServerMetadata.options.maxAge = 1000 * serverConfig.AUTH_RAOIDC_METADATA_CACHE_TTL_SECONDS;
