+++ frontend/app/.server/utils/raoidc.utils.ts
@@ -7,7 +7,7 @@ import { SignJWT, compactDecrypt, importJWK, jwtVerify } from 'jose';
 import { createHash, subtle } from 'node:crypto';
 import invariant from 'tiny-invariant';
 
-import type { FetchFn } from '~/.server/utils/fetch.utils';
+import type { FetchFn } from '~/.server/http';
 import { getLogger } from '~/.server/utils/logging.utils';
 
 /**
